Finalized sidebar style file cleanup
